Extracted nested classes from ViewUtils
