+++ src/main/java/it/unibz/inf/ontouml/vp/utils/ViewUtils.java
@@ -666,98 +666,3 @@ public class ViewUtils {
 	}
 }
 
-@SuppressWarnings("serial")
-final class ContextMenu extends JPopupMenu {
-	private JMenuItem takeMeThere;
-	private JMenuItem openSpec;
-	private ActionListener menuListener;
-
-	public ContextMenu() {
-
-	}
-
-	public ContextMenu(String idModelElement) {
-
-		menuListener = new ActionListener() {
-			@Override
-			public void actionPerformed(ActionEvent e) {
-				switch (e.getActionCommand()) {
-				case "Take me there!":
-					System.out.println("Firing 'Highlight'");
-					ViewUtils.highlightDiagramElement(idModelElement);
-					break;
-				case "Open Specification":
-					System.out.println("Firing 'Open Specification'");
-					ViewUtils.openSpecDiagramElement(idModelElement);
-					break;
-				default:
-					break;
-				}
-			}
-		};
-
-		takeMeThere = new JMenuItem("Take me there!", new ImageIcon(ViewUtils.getFilePath(ViewUtils.NAVIGATION_LOGO)));
-		takeMeThere.addActionListener(menuListener);
-		openSpec = new JMenuItem("Open Specification", new ImageIcon(ViewUtils.getFilePath(ViewUtils.MORE_HORIZ_LOGO)));
-		openSpec.addActionListener(menuListener);
-		add(takeMeThere);
-		add(openSpec);
-	}
-
-	public void disableItem(String item) {
-		switch (item) {
-		case "Take me there!":
-			takeMeThere.setEnabled(false);
-			break;
-		case "Open Specification":
-			openSpec.setEnabled(false);
-			break;
-		default:
-			break;
-		}
-	}
-}
-
-final class ContextMenuListener extends MouseAdapter {
-	private ArrayList<String> idModelElementList;
-	private JList<Object> messageList;
-
-	ContextMenuListener(ArrayList<String> list, JList<Object> messages) {
-		super();
-		idModelElementList = list;
-		messageList = messages;
-	}
-
-	@Override
-	public void mouseMoved(MouseEvent e) {
-		final Point p = e.getPoint();
-		final int index = messageList.locationToIndex(p);
-
-		messageList.setSelectedIndex(index);
-	}
-
-	@Override
-	public void mouseExited(MouseEvent e) {
-		messageList.clearSelection();
-	}
-
-	public void mouseReleased(MouseEvent e) {
-		doPop(e);
-	}
-
-	private void doPop(MouseEvent e) {
-		ContextMenu menu;
-		String idModelElement = idModelElementList.get(messageList.locationToIndex(e.getPoint()));
-
-		if (idModelElement == null) {
-			menu = new ContextMenu();
-		} else {
-			menu = new ContextMenu(idModelElement);
-			if (!ViewUtils.isElementInAnyDiagram(idModelElement)) {
-				menu.disableItem("Take me there!");
-			}
-		}
-
-		menu.show(e.getComponent(), e.getX(), e.getY());
-	}
-}
